+++ contao/dca/tl_page.php
@@ -27,7 +27,6 @@ PaletteManipulator::create()
     ->addLegend('schema_legend', 'meta_legend', PaletteManipulator::POSITION_AFTER)
     ->addField('headAddOrganisationSchema', 'schema_legend', PaletteManipulator::POSITION_APPEND)
     ->addField('headAddWebSiteSchema', 'schema_legend', PaletteManipulator::POSITION_APPEND)
-    ->addField('headAddBreadcrumbSchema', 'schema_legend', PaletteManipulator::POSITION_APPEND)
     ->applyToPalette('root', 'tl_page')
     ->applyToPalette('rootfallback', 'tl_page');
 
@@ -99,14 +98,6 @@ $fields = [
         'sql' => 'binary(16) NULL',
     ],
     'headAddWebSiteSchema' => [
-        'exclude' => true,
-        'inputType' => 'checkbox',
-        'eval' => [
-            'tl_class' => 'w50 clr',
-        ],
-        'sql' => "char(1) NOT NULL default ''",
-    ],
-    'headAddBreadcrumbSchema' => [
         'exclude' => true,
         'inputType' => 'checkbox',
         'eval' => [